Removes mocha dependecy from SqliteDBCreateTest
Rewrite test using Minitest and Minitest::Mocks only, and since it does
not allow to set expectations on any Object other than `Minitest::Mock`,
decided to invert control on the database tasks, so that both the
dependecies to the file system and the database are now passed in as a
configuration.

In order to completely mimic what was already being done with mocha,
there is a need to count the number of times a method is called in a
minitest mock (in this case it is 0 times). This is a feature that has
been discussed on minitest already, so we might just add it sooner than
later. Reference https://github.com/seattlerb/minitest/pull/566 for more
on the discussion.